% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{cryptocurrency_list}
\alias{cryptocurrency_list}
\title{Coinmarketcap Cryptocurrency Database}
\format{A tidy data frame with 4 columns}
\source{
\url{https://coinmarketcap.com}
}
\usage{
cryptocurrency_list
}
\description{
A dataset containing the name and symbology of all active currencies at coinmarketcap
}
\keyword{datasets}
