% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_data.R
\name{historical_data}
\alias{historical_data}
\title{Returns a Tidy data frame of historical data for selected cryptocurrencies}
\usage{
historical_data(cryptocurrency = c("BTC", "PPC"))
}
\arguments{
\item{cryptocurrency}{A vector with cryptocurrency symbols}
}
\value{
A data frame with fourteen columns.
}
\description{
Returns a tidy data frame of historical data for specificed cryptocurrencies
}
\details{
The tidy data frame constains general information about the selected cryptocurrencies on date.
}
\examples{

data <- historical_data(cryptocurrency = c('BTC'))


}
