% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listing.R
\name{listing}
\alias{listing}
\title{Tidy data frame of all active cryptocurrency at coinmarketcap}
\usage{
listing()
}
\value{
A data frame with four columns: \code{id}, \code{name}, \code{symbol} and \code{website_slug},
}
\description{
Returns a tidy data frame of all active cryptocurrency at coinmarketcap
}
\details{
The tidy data frame constains meta data about all active cryptocurrency,
including (1) all active cryptocurrency´s id, (2) cryptocurrency name, (3) symbol, and the (4)
correspondent website slug.
}
\examples{

coins <- listing()


}
