% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrive_data.R
\name{retrive_data}
\alias{retrive_data}
\title{Webscraping historical data for selected cryptocurrencies}
\usage{
retrive_data(id)
}
\arguments{
\item{id}{A vector with cryptocurrency slug name}
}
\value{
A list.
}
\description{
Scraps historical data for specificed cryptocurrencies
}
\details{
The tidy data frame contains historical data for selected cryptocurrencies in list format.
}
\examples{

data <- retrive_data(id = c('bitcoin'))


}
