% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticker.R
\name{ticker}
\alias{ticker}
\title{Tidy data frame of ticker data for specificed cryptocurrencies}
\usage{
ticker(cryptocurrency = c("BTC"))
}
\arguments{
\item{cryptocurrency}{A vector with cryptocurrencies symbol}
}
\value{
A data frame with fourteen columns.
}
\description{
Returns a tidy data frame of ticker data for specificed cryptocurrencies
}
\details{
The tidy data frame constains general information about the selected cryptocurrencies on date.
}
\examples{

ticker_data <- ticker(cryptocurrency = c('BTC', 'PPC'))


}
